refactor(annotation): one timezone-aware datepicker for labels and primitives
The annotate sidebar had two date edit paths: sample primitives used an
inline voodo DatePicker with timezone-aware serialization, while label
attributes went through the SmartForm DatePickerWidget with none — so
label date fields still rendered the stored UTC instant in the browser
timezone.

Move the picker/timezone conversion helpers into @fiftyone/utilities
next to the existing timezone helpers, make DatePickerWidget convert
with them (UTC calendar date for date fields, fo.config.timezone wall
clock for datetimes), and drop the PrimitiveRenderer date special-case
so both surfaces share the widget. Form data now carries dates as ISO
instant strings on both paths.

Also declare the pre-existing luxon dependency in @fiftyone/utilities.
