+++ app/packages/core/src/components/Modal/Sidebar/Annotate/Edit/PrimitiveRenderer.tsx
@@ -1,6 +1,5 @@
 import type { SchemaType } from "@fiftyone/core/src/plugins/SchemaIO/utils/types";
 import { Primitive } from "@fiftyone/utilities";
-import { DatePicker } from "@voxel51/voodo";
 import styled from "styled-components";
 import { SchemaIOComponent } from "../../../../../plugins/SchemaIO";
 import JSONEditor, {
@@ -15,7 +14,7 @@ const EditorContainer = styled.div`
 
 interface PrimitiveRendererProps {
   type: string;
-  fieldValue: Primitive | Date;
+  fieldValue: Primitive;
   handleChange: (data: unknown) => void;
   primitiveSchema: SchemaType | undefined;
 }
@@ -27,7 +26,6 @@ export default function PrimitiveRenderer({
   primitiveSchema,
 }: PrimitiveRendererProps) {
   const isJson = type === "dict";
-  const isDate = type === "date" || type === "datetime";
   if (isJson) {
     return (
       <EditorContainer>
@@ -39,22 +37,6 @@ export default function PrimitiveRenderer({
           showDocumentation={false}
         />
       </EditorContainer>
-    );
-  }
-
-  if (isDate) {
-    return (
-      <DatePicker
-        selected={fieldValue as Date}
-        showTimeSelect={type === "datetime"}
-        onChange={(date: Date | null) => {
-          if (date && !Number.isNaN(date.getTime())) {
-            handleChange(date.toISOString());
-          } else {
-            handleChange(undefined);
-          }
-        }}
-      />
     );
   }
