+++ app/packages/utilities/package.json
@@ -29,6 +29,7 @@
         "fetch-retry": "^5.0.3",
         "lodash": "^4.18.1",
         "lru-cache": "^11.0.1",
+        "luxon": "^3.6.1",
         "mime": "^2.5.2",
         "path-browserify": "^1.0.1"
     },
