+++ app/packages/utilities/src/datetime.ts
@@ -84,3 +84,95 @@ const handleDigits = (digits: string) => {
     useGrouping: false,
   });
 };
+
+/**
+ * FiftyOne renders `date` fields as the UTC calendar date and `datetime`
+ * fields in the app timezone (`fo.config.timezone`, "UTC" by default),
+ * while react-datepicker only operates on Date objects interpreted in the
+ * browser's local timezone. The helpers below translate between the two so
+ * pickers show and store the same wall-clock values the rest of the app
+ * displays.
+ */
+
+/**
+ * Convert an absolute timestamp into a Date whose local-timezone components
+ * match what the app displays for the field, suitable for react-datepicker
+ * @param type - the type of the field ("date" or "datetime")
+ * @param timestamp - epoch milliseconds
+ * @param timeZone - the app display timezone (IANA name, "UTC", or "local")
+ * @returns a Date carrying the displayed wall-clock values in local time
+ */
+export function toPickerDate(
+  type: string,
+  timestamp: number,
+  timeZone: string,
+): Date {
+  if (type === "date") {
+    const date = new Date(timestamp);
+    return new Date(
+      date.getUTCFullYear(),
+      date.getUTCMonth(),
+      date.getUTCDate(),
+      12,
+    );
+  }
+
+  const dt = DateTime.fromMillis(timestamp, { zone: timeZone });
+  return new Date(
+    dt.year,
+    dt.month - 1,
+    dt.day,
+    dt.hour,
+    dt.minute,
+    dt.second,
+    dt.millisecond,
+  );
+}
+
+/**
+ * Serialize a picker Date's calendar date to an absolute instant at noon
+ * UTC, so the stored value renders as the same calendar date in every
+ * timezone
+ */
+export function dateOnlyToUTC(date: Date): string {
+  const year = date.getFullYear();
+  const month = date.getMonth();
+  const day = date.getDate();
+
+  return new Date(Date.UTC(year, month, day, 12, 0, 0, 0)).toISOString();
+}
+
+export function serializeDateValue(
+  type: string,
+  date: Date,
+  timeZone: string,
+): string {
+  if (type === "date") {
+    return dateOnlyToUTC(date);
+  }
+
+  // the picker's Date components are wall-clock values in the app display
+  // timezone; interpret them there to recover the absolute instant
+  const iso = DateTime.fromObject(
+    {
+      year: date.getFullYear(),
+      month: date.getMonth() + 1,
+      day: date.getDate(),
+      hour: date.getHours(),
+      minute: date.getMinutes(),
+      second: date.getSeconds(),
+      millisecond: date.getMilliseconds(),
+    },
+    { zone: timeZone },
+  )
+    .toUTC()
+    .toISO();
+
+  // an invalid zone or date yields null; throw so callers skip the mutation
+  // instead of treating the value as missing and deleting the field
+  if (iso === null) {
+    throw new Error(`invalid date or timezone: ${date} (${timeZone})`);
+  }
+
+  return iso;
+}
